requirements.txt
anthropic>=0.40.0
slack-bolt>=1.20.0
slack-sdk>=3.30.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
